requirements.txt
alembic==1.8.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
python-dateutil==2.8.2
SQLAlchemy==1.4.37
passlib==1.7.4
psycopg2-binary==2.8.6
PyJWT==2.6.0
